fix: to fix error from write_h5ad for lfq data, change psm.X datatype to float in _sage2mdata in LfqSageReader

+++ msmu/_read_write/_sage_reader.py
@@ -254,7 +254,7 @@ class LfqSageReader(SageReader):
         sage_quant_df = sage_quant_df.rename(columns=rename_dict)
 
         adata_psm = ad.AnnData(
-            pd.DataFrame(index=sage_result_df.index, columns=sage_quant_df.columns).T
+            pd.DataFrame(index=sage_result_df.index, columns=sage_quant_df.columns).T.astype("float")
         )
         adata_psm.var = self._normalise_columns(sage_result_df, self._precursor_charge)
         adata_psm.varm["search_result"] = sage_result_df
